todo/src/requirements.txt
amqp==5.0.9
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==4.0.2
django-crispy-forms==1.14.0
fpdf==1.7.2
kombu==5.2.3
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.26
psycopg2==2.9.3
pyparsing==3.0.7
pytz==2021.3
redis==4.1.3
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
utmp==21.10.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
